s_binMonitoring/requirements.txt
async-timeout==5.0.1
attrs==25.1.0
blinker==1.9.0
click==8.1.8
flasgger==0.9.7.1
Flask==3.1.0
flask-cors==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kafka-python==2.1.4
MarkupSafe==3.0.2
mistune==3.1.2
packaging==24.2
paho-mqtt==2.1.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.23.1
six==1.17.0
typing_extensions==4.12.2
Werkzeug==3.1.3
